requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpatch==1.28
jsonpointer==2.0
MarkupSafe==1.1.1
PyYAML==5.4.1
Werkzeug==1.0.1
